+++ src/components/CharacterListing/CharacterListingContext.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, useEffect, useReducer, useState } from "react";
+import React, { createContext, useEffect, useReducer } from "react";
 import { Character } from "../../types";
 
 // @todo: This should be suffixed with a const assertion, but create-react-app
